% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing-risks.R
\name{cuminc_naive}
\alias{cuminc_naive}
\title{Naive (unweighted) cumulative incidence estimate}
\usage{
cuminc_naive(dat, esttimes)
}
\arguments{
\item{dat}{A wide-format competing risks data frame with columns \code{t} and
\code{delta} (factor with levels \code{"censor"}, \code{"event_1"}, \code{"event_2"}).}

\item{esttimes}{Numeric vector of times at which to return estimates.}
}
\value{
A numeric vector of cumulative incidence estimates at \code{esttimes}.
}
\description{
Estimates the cumulative incidence of event type 1 using the standard
Aalen-Johansen estimator without any IPCW adjustment. Serves as a
comparison method when censoring is independent of covariates.
}
\examples{
set.seed(42)
dat <- sim_data_CR(n = 200, censoring = "independent")
cuminc_naive(dat, esttimes = seq(0, 5, 0.5))

}
