% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing-risks.R
\name{cuminc_naive_cr}
\alias{cuminc_naive_cr}
\title{Naive (unweighted) cumulative incidence estimate}
\usage{
cuminc_naive_cr(dat, esttimes, time_var = "t", event_var = "delta")
}
\arguments{
\item{dat}{A wide-format competing risks data frame containing the columns
specified by \code{time_var} and \code{event_var}.}

\item{esttimes}{Numeric vector of times at which to return estimates.}

\item{time_var}{Character string. Name of the event/censoring time column.
Default is \code{"t"}.}

\item{event_var}{Character string. Name of the event indicator column.
Default is \code{"delta"}.}
}
\value{
A numeric vector of cumulative incidence estimates at \code{esttimes}.
}
\description{
Estimates the cumulative incidence of event type 1 using the standard
Aalen-Johansen estimator without any IPCW adjustment. Serves as a
comparison method when censoring is independent of covariates.
}
\examples{
set.seed(42)
dat <- sim_data_cr(n = 200, censoring = "independent")
cuminc_naive_cr(dat, esttimes = seq(0, 5, 0.5))

}
