% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-event.R
\name{get_boot_var}
\alias{get_boot_var}
\title{Compute bootstrap variance of the log hazard ratio}
\usage{
get_boot_var(data, B)
}
\arguments{
\item{data}{A data frame with a column \code{log_hr} containing bootstrap
log hazard ratio estimates, as returned by multiple calls to
\code{\link[=get_ipcw_cox_fit]{get_ipcw_cox_fit()}} or \code{\link[=get_cox_fit]{get_cox_fit()}}.}

\item{B}{Integer. Number of bootstrap samples used.}
}
\value{
A single numeric value: the bootstrap variance estimate.
}
\description{
Compute bootstrap variance of the log hazard ratio
}
\examples{
# Toy example with 10 simulated bootstrap log HRs
boot_results <- data.frame(log_hr = rnorm(10, mean = -0.5, sd = 0.2))
get_boot_var(boot_results, B = 10)

}
